split index into many files
