+++ packages/@coorpacademy-app-player/src/view/popin-end.js
@@ -1,0 +1,9 @@
+import PopinEnd from '@coorpacademy/components/es/template/app-player/popin-end';
+import mapStateToProps from './state-to-props/popin-end';
+import createView from './create-view';
+
+const createPopinEnd = (options, dispatch) => {
+  return createView(options, dispatch, PopinEnd, mapStateToProps);
+};
+
+export default createPopinEnd;
