+++ packages/@coorpacademy-app-player/src/view/test/answer-props.js
@@ -1,7 +1,7 @@
 import test from 'ava';
 import isFunction from 'lodash/fp/isFunction';
 import reduce from 'lodash/fp/reduce';
-import getAnswerProps from '../answer-props';
+import getAnswerProps from '../state-to-props/answer';
 import {ANSWER_EDIT} from '../../actions/ui/answers';
 import basic from './fixtures/slides/basic';
 import qcm from './fixtures/slides/qcm';
