+++ packages/@coorpacademy-app-player/src/view/test/index.js
@@ -2,6 +2,6 @@ import test from 'ava';
 import isFunction from 'lodash/fp/isFunction';
 import {create} from '../..';
 
-test('app-player shoud provide a create function', t => {
+test('app-player should provide a create function', t => {
   t.is(isFunction(create), true);
 });
