+++ packages/@coorpacademy-app-player/src/view/test/init.js
@@ -1,9 +1,9 @@
 import noop from 'lodash/fp/noop';
 import test from 'ava';
-import {createMapStateToProps} from '..';
+import {slideStateToProps} from '../slide';
 import state from './fixtures/init-state';
 
-const mapStateToProps = createMapStateToProps({})(noop);
+const mapStateToProps = slideStateToProps({}, noop);
 
 test('should provide default props to init state', t => {
   const props = mapStateToProps(state);
